* objc/sarray.h: Make boffset be an unsigned long when sparc so it 	works out on 64-bit systems.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@22881 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libobjc/objc/sarray.h
@@ -92,7 +92,7 @@ struct soffset {
   unsigned int ioffset : SIZET_BITS/4;
 #else /* OBJC_SPARSE2 */
 #ifdef __sparc__
-  unsigned int boffset : (SIZET_BITS - 2) - BUCKET_BITS;
+  unsigned long boffset : (SIZET_BITS - 2) - BUCKET_BITS;
   unsigned int eoffset : BUCKET_BITS;
   unsigned int unused  : 2;
 #else
